chore: move method to utils
