+++ express/blocks/app-store-highlight/app-store-highlight.js
@@ -12,8 +12,9 @@
 
 import {
   createOptimizedPicture, createTag, fetchPlaceholders, getIcon, getIconElement, getMetadata,
-  getMobileOperatingSystem,
 } from '../../scripts/scripts.js';
+
+import { getMobileOperatingSystem } from '../shared/utils.js';
 
 const imageSrcs = [
   '/express/media_1e71d9d4a13e8a8422dd1b8dbdad9a0bf4d2565f8.jpeg?width=380&format=jpeg&optimize=medium',
